+++ x/leverage/keeper/collateral.go
@@ -7,6 +7,40 @@ import (
 	"github.com/umee-network/umee/v2/x/leverage/types"
 )
 
+// liquidateCollateral burns uToken collateral and sends the base token reward to the liquidator.
+// This occurs during direct liquidation.
+func (k Keeper) liquidateCollateral(ctx sdk.Context, borrower, liquidator sdk.AccAddress, uToken, token sdk.Coin,
+) error {
+	if err := k.burnCollateral(ctx, borrower, uToken); err != nil {
+		return err
+	}
+	return k.bankKeeper.SendCoinsFromModuleToAccount(ctx, types.ModuleName, liquidator, sdk.NewCoins(token))
+}
+
+// burnCollateral removes some uTokens from an account's collateral and burns them. This occurs
+// during liquidations.
+func (k Keeper) burnCollateral(ctx sdk.Context, addr sdk.AccAddress, uToken sdk.Coin) error {
+	err := k.setCollateralAmount(ctx, addr, k.GetCollateralAmount(ctx, addr, uToken.Denom).Sub(uToken))
+	if err != nil {
+		return err
+	}
+	if err = k.bankKeeper.BurnCoins(ctx, types.ModuleName, sdk.NewCoins(uToken)); err != nil {
+		return err
+	}
+	return k.setUTokenSupply(ctx, k.GetUTokenSupply(ctx, uToken.Denom).Sub(uToken))
+}
+
+// decollateralize removes fromAddr's uTokens from the module and sends them to toAddr.
+// It occurs when decollateralizing uTokens (in which case fromAddr and toAddr are the
+// same) as well as during non-direct liquidations, where toAddr is the liquidator.
+func (k Keeper) decollateralize(ctx sdk.Context, fromAddr, toAddr sdk.AccAddress, uToken sdk.Coin) error {
+	err := k.setCollateralAmount(ctx, fromAddr, k.GetCollateralAmount(ctx, fromAddr, uToken.Denom).Sub(uToken))
+	if err != nil {
+		return err
+	}
+	return k.bankKeeper.SendCoinsFromModuleToAccount(ctx, types.ModuleName, toAddr, sdk.NewCoins(uToken))
+}
+
 // GetCollateralAmount returns an sdk.Coin representing how much of a given denom the
 // x/leverage module account currently holds as collateral for a given borrower.
 func (k Keeper) GetCollateralAmount(ctx sdk.Context, borrowerAddr sdk.AccAddress, denom string) sdk.Coin {
@@ -54,10 +88,10 @@ func (k Keeper) setCollateralAmount(ctx sdk.Context, borrowerAddr sdk.AccAddress
 }
 
 // GetTotalCollateral returns an sdk.Coin representing how much of a given uToken
-// the x/leverage module account currently holds as collateral. Non-uTokens and invalid
-// assets return zero.
+// the x/leverage module account currently holds as collateral. Non-uTokens return
+// zero.
 func (k Keeper) GetTotalCollateral(ctx sdk.Context, denom string) sdk.Int {
-	if !k.IsAcceptedUToken(ctx, denom) {
+	if !types.HasUTokenPrefix(denom) {
 		// non-uTokens cannot be collateral
 		return sdk.ZeroInt()
 	}