+++ x/leverage/types/errors.go
@@ -38,4 +38,6 @@ var (
 		1126,
 		"market total collateral would exceed MaxCollateralShare",
 	)
+	ErrNotUToken = sdkerrors.Register(ModuleName, 1127, "denom should be a uToken")
+	ErrUToken    = sdkerrors.Register(ModuleName, 1128, "denom should not be a uToken")
 )